+++ IconList.lua
@@ -1,121 +1,54 @@
 --[[ 
     Gma3 Icon List
-    version 1.0.0
+    version 1.1.0
     Last update April 16, 2025
     Developed by Jeremy Dufeux - Carrot Industries
     https://carrot-industries.com
     contact: contact@carrot-industries.com
 
-    License described in the LICENSE.txt file
+    This software is licensed as described in the LICENSE.txt file.
 ]]
 
+local signalTable = select(3, ...)
 local pluginHandle = select(4, ...)
 
-Icon = {
-  arrow_down = "arrow_down",
-  arrow_up = "arrow_up",
-  backup = "backup",
-  cooking = "cooking",
-  execconfig = "execconfig",
-  FlipLeftRight = "FlipLeftRight",
-  icon_timer = "icon_timer",
-  minus = "minus",
-  object_3d = "object_3d",
-  object_agenda = "object_agenda",
-  object_align = "object_align",
-  object_appear = "object_appear",
-  object_bitmaps = "object_bitmaps",
-  object_camera = "object_camera",
-  object_clock = "object_clock",
-  object_cmd = "object_cmd",
-  object_colpick = "object_colpick",
-  object_content = "object_content",
-  object_commandwing = "object_commandwing",
-  object_datapool = "object_datapool",
-  object_dmx = "object_dmx",
-  object_encbar = "object_encbar",
-  object_filter = "object_filter",
-  object_fixture = "object_fixture",
-  object_fps = "object_fps",
-  object_gels = "object_gels",
-  object_generators = "object_generators",
-  object_group1 = "object_group1",
-  object_gobos = "object_gobos",
-  object_layout = "object_layout",
-  object_info = "object_info",
-  object_images = "object_images",
-  object_macro = "object_macro",
-  object_matricks = "object_matricks",
-  object_meshes = "object_meshes",
-  object_message = "object_message",
-  object_pages = "object_pages",
-  object_playback = "object_playback",
-  object_plugin1 = "object_plugin1",
-  object_preset = "object_preset",
-  object_phase = "object_phase",
-  object_question_small = "object_question_small",
-  object_quickey = "object_quickey",
-  object_runningpb = "object_runningpb",
-  object_scribbles = "object_scribbles",
-  object_sel_bar = "object_sel_bar",
-  object_selgrid = "object_selgrid",
-  object_sequence = "object_sequence",
-  object_smart = "object_smart",
-  object_sound3 = "object_sound3",
-  object_special_master = "object_special_master",
-  object_stepbar = "object_stepbar",
-  object_symbols = "object_symbols", 
-  object_sysinfo = "object_sysinfo",
-  object_sysmonitor = "object_sysmonitor",
-  object_tags = "object_tags",
-  object_timecode = "object_timecode",
-  object_timecodeslots = "object_timecodeslots",
-  object_trackp = "object_trackp",
-  object_universe = "object_universe",
-  object_user = "object_user",
-  object_video = "object_video",
-  object_view = "object_view",
-  object_world1 = "object_world1",
-  object_xkeys = "object_xkeys",
-  PhaserCopy = "PhaserCopy",
-  PhaserCut = "PhaserCut",
-  PhaserPaste = "PhaserPaste",
-  plus = "plus",
-  star = "star",
-  time = "time",
-  tools = "tools",
-}
-
-function Main
-  ()
-  local contentRoot = CreateInputDialog()
-
-  local columns = 4
-  local rows= 12
-
-  local grid = contentRoot:Append("UILayoutGrid")
-  grid.Columns = columns
-  grid.Rows = rows
-
-  local actualColumns = 0
-  local actualRows = 0
-
-  for _, icon in pairs(Icon) do
-    AddButton(grid, actualColumns, actualRows, icon, icon)
-    if(actualColumns == columns) then
-      actualColumns = 0
-      actualRows = actualRows + 1
-    else
-      actualColumns = actualColumns + 1
-    end
-  end
+if IconList == nil then
+  IconList = {}
 end
 
-function CreateInputDialog()
+function IconList.Main()
   local displayIndex = Obj.Index(GetFocusDisplay())
-   if displayIndex > 5 then
-     displayIndex = 1
-   end
+  if displayIndex > 5 then
+    displayIndex = 1
+  end
+  
+  local icons = IconList.CreateIconTableFromXMLFile()
+  if(icons == nil) then
+    Printf("No icons found")
+    return
+  end
+
+  local width = 1600
+  local height = 800
+
+  local columns = 4
+  local rows= 50
+
+  local buttonSize = 120
+
+  local iconsAmount = IconList.GetTableSize(icons)
+  if(iconsAmount == 0) then
+    Printf("No icons found")
+    return
+  end
+
+  if iconsAmount % columns > 0 then
+    rows = math.floor(iconsAmount / columns) + 1
+  else
+    rows = math.floor(iconsAmount / columns)
+  end
+  local scrollHeight = rows * buttonSize
+
   local display = GetDisplayByIndex(displayIndex)
   local screenOverlay = display.ScreenOverlay
   screenOverlay:ClearUIChildren()
@@ -123,7 +56,7 @@ function CreateInputDialog()
   local baseInput = screenOverlay:Append("BaseInput")
   baseInput.Name = "Gma3 Icons list"
   baseInput.H = "0"
-  baseInput.W = 1200 
+  baseInput.W = width 
   baseInput.Rows = 2
   baseInput[1][1].SizePolicy = "Fixed"
   baseInput[1][1].Size = "60"
@@ -137,50 +70,179 @@ function CreateInputDialog()
   titleBar.Texture = "corner2"
 
   local titleBarIcon = titleBar:Append("TitleButton")
-  titleBarIcon.Text = "Gma3 Icons list"
+  titleBarIcon.Text = "Gma3 Icons list - "..iconsAmount.." icons"
   titleBarIcon.Texture = "corner1"
   titleBarIcon.Anchors = "0,0"
-  titleBarIcon.Icon = Icon.object_sequence
+  titleBarIcon.Icon = "object_sequence"
 
   local titleBarCloseButton = titleBar:Append("CloseButton")
   titleBarCloseButton.Anchors = "1,0"
   titleBarCloseButton.Texture = "corner2"
 
-  local contentRoot = baseInput:Append("DialogFrame")
-  contentRoot.Anchors = {
-    left = 0,
-    right = 0,
-    top = 1,
-    bottom = 1
-  }
-  contentRoot[1][1].SizePolicy = "Fixed"
-  contentRoot[1][1].Size = 1000
+  local dlgFrame = baseInput:Append("DialogFrame")
+	dlgFrame.H = "100%"
+	dlgFrame.W = "100%"
+	dlgFrame.Columns = 2
+	dlgFrame.Rows = 1
+	dlgFrame.Anchors = { left = 0, right = 0, top = 1, bottom = 1 }
 
-  return contentRoot
+	dlgFrame[1][1].SizePolicy = "Content"
+	dlgFrame[2][1].SizePolicy = "Stretch"
+	dlgFrame[2][2].SizePolicy = "Content"
+
+	local boxHolder = dlgFrame:Append("UIObject")
+	boxHolder.Anchors = { left = 0, right = 1, top = 0, bottom = 0 }
+	boxHolder.H, boxHolder.W = height, "100%"
+	boxHolder.Transparent = true
+
+	local scrollbox = boxHolder:Append("ScrollBox")
+
+	local scrollbar = dlgFrame:Append("ScrollBarV")
+	scrollbar.ScrollTarget = scrollbox
+	scrollbar.Anchors = "1,0"
+
+	local grid = scrollbox:Append("UILayoutGrid")
+	grid.W, grid.H = width-48, scrollHeight
+  grid.Columns = columns
+  grid.Rows = rows
+
+  local actualColumns = 0
+  local actualRows = 0
+
+  for _, icon in pairs(IconList.GetSortedKeys(icons)) do
+    IconList.AddButton(grid, actualColumns, actualRows, icon, icon, buttonSize)
+    if(actualColumns == columns-1) then
+      actualColumns = 0
+      actualRows = actualRows + 1
+    else
+      actualColumns = actualColumns + 1
+    end
+  end
 end
 
-function AddButton(parent, x, y, text, icon)
-  local button = parent:Append("Button")
-  button.PluginComponent = pluginHandle
-  button.Text = text
-  button.Font = "Medium20"
-  button.TextalignmentH = "Left"
-  button.Textshadow = 1;
-  button.Anchors = {
+function IconList.CreateIconTableFromXMLFile(xmlFilePath)
+  local texturePath = GetPath("textures")
+  local textureFilePath = texturePath.."/graphics.textures.xml"
+  
+  local file = io.open(textureFilePath, "r")
+  if not file then
+      Printf("Can't find the graphics.textures.xml file")
+      Printf("It should be located in "..texturePath)
+      return nil
+  end
+  
+  local content = file:read("*all")
+  file:close()
+  
+  local cleanedContent = content:gsub("<!%-%-.-%-%->"," ")
+  
+  local icons = {}
+  for name in cleanedContent:gmatch("<Texture%s+Name=\"([^\"]+)\"") do
+      icons[name] = name
+  end
+  
+  return icons
+end
+
+function IconList.GetSortedKeys(tbl)
+  local keys = {}
+
+  for key in pairs(tbl) do
+    table.insert(keys, key)
+  end
+  
+  table.sort(keys, function(a, b)
+    return string.lower(a) < string.lower(b)
+  end)
+
+  return keys
+end
+
+function IconList.GetTableSize(t)
+  local count = 0
+  for _ in pairs(t) do
+      count = count + 1
+  end
+  return count
+end
+
+function IconList.CopyToClipboard(text)
+  local tmpFile = os.tmpname()
+  local file = io.open(tmpFile, "w")
+  
+  if not file then
+    Printf("Failed to create temporary file")
+    return
+  end
+
+  file:write(text)
+  file:close()
+  
+  local host = HostOS()
+
+  if(host == "Windows") then
+    os.execute("type " .. tmpFile .. " | clip")
+  elseif(host == "Linux") then
+    os.execute("cat " .. tmpFile .. " | xclip -selection clipboard")
+  elseif(host == "Mac") then
+    os.execute("cat " .. tmpFile .. " | pbcopy")
+  else
+    Printf("Unsupported OS: " .. host)
+  end
+end
+
+function IconList.AddButton(parent, x, y, text, icon, buttonSize)
+  local margin = 4
+
+  local grid = parent:Append("UILayoutGrid")
+  grid.Anchors = {
     left = x,
     right = x,
     top = y,
     bottom = y
   }
-  button.Margin = {
-    left = 2,
-    right = 2,
-    top = 2,
-    bottom = 2
+  grid.Columns = 2
+  grid.Rows = 1
+  grid.Margin = {
+    left = margin*2,
+    right = margin*2,
+    top = margin,
+    bottom = margin
   }
-  button.Padding = "5,5"
-  button.IconAlignmentH = "Right"
+  grid.W = 520
+
+  local label = grid:Append("Button")
+  label.PluginComponent = pluginHandle
+  label.Text = text
+  label.Font = "Medium20"
+  label.TextalignmentH = "Left"
+  label.Textshadow = 1;
+  label.Anchors = {
+    left = 0,
+    right = 0,
+    top = 0,
+    bottom = 0
+  }
+  label.Padding = "5,5"
+  label.Clicked = "Callback"
+
+  local button = grid:Append("Button")
+  button.PluginComponent = pluginHandle
+  button.Anchors = {
+    left = 1,
+    right = 1,
+    top = 0,
+    bottom = 0
+  }
   button.Icon = icon
+  button.W = buttonSize
+  button.Clicked = "Callback"
+
+  signalTable.Callback = function(caller)
+    Printf("Clicked on "..caller.Text)    
+    IconList.CopyToClipboard(caller.Text)
+  end
+
 end
 
-return Main
+return IconList.Main