+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Change Log
 
+## 1.1.0 — April 16, 2025
+
+- Add scroll view in the dialog window
+- Change icons source to get them directly in the graphics.textures.xml file
+- Add copy to clipboard function when click on icons
+- Update LICENSE.TXT file
+- Update README.md file
+
 ## 1.0.0 — April 16, 2025
+
 - Initial public release